fix bug val leastDeposit: Double? = null,
it was
val leastDeposit: Int? = null,

+++ data/src/main/java/com/example/data/model/models/ModelsDto.kt
@@ -24,7 +24,7 @@ data class ModelsDto(
 	val hiddenPrice: Int? = null,
 
 	@field:SerializedName("least_deposit")
-	val leastDeposit: Int? = null,
+	val leastDeposit: Double? = null,
 
 	@field:SerializedName("attributes_hint_list")
 	val attributesHintList: List<AttributesHintListItem?>? = null,
